+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Creating patch for SvqParser and test
 * New Trip class with RSpec
 * More Segments classes and RSpecs: Train, Flight and Hotel
 * Fix specs and refactoring